Fix issue #1: Not all packages are shown.
Because existing_package is an sqlalchemy.orm.query.Query object it
evaluates to True. So even if there were no results, the code would
not create a new Package. (And, since the first result of the Query
would be None, the packages_list would have None appended instead of a
Package.)